requirements.txt
sanic==24.12.0
sanic-jwt==1.8.0
python-dotenv==1.0.1
aiogram==3.17.0
aiohttp==3.11.11
sqlalchemy==2.0.38
greenlet==3.1.1
asyncpg==0.30.0
ujson==5.10.0
pyjwt==2.10.1
bcrypt==4.2.1
